+++ tests/test_sqlalchemy_registry.py
@@ -8,9 +8,10 @@ import sqlalchemy as sa  # noqa: F401  # pylint: disable=unused-import
 
 import pytest
 
-from coaster.db import db
 from coaster.sqlalchemy import BaseMixin
 from coaster.sqlalchemy.registry import Registry
+
+from .test_sqlalchemy_models import db
 
 # --- Fixtures -------------------------------------------------------------------------
 
@@ -113,13 +114,13 @@ def registrymixin_models():
     # in the base RegistryMixin class.
 
     # Sample model 1
-    class RegistryTest1(BaseMixin, db.Model):
+    class RegistryTest1(BaseMixin, db.Model):  # type: ignore[name-defined]
         """Registry test model 1."""
 
         __tablename__ = 'registry_test1'
 
     # Sample model 2
-    class RegistryTest2(BaseMixin, db.Model):
+    class RegistryTest2(BaseMixin, db.Model):  # type: ignore[name-defined]
         """Registry test model 2."""
 
         __tablename__ = 'registry_test2'